+++ game/src/components/Index/FleetStatus.tsx
@@ -1,6 +1,6 @@
 import { api } from "@/utils/api";
 
-import LoadingSpinner from "../Loader/LoadingSpinner";
+import LoadingSpinner from "@/components/common/Loader/LoadingSpinner";
 
 interface FleetStatusProps {
   paPlayer: {
