+++ game/src/components/Index/LandTable.tsx
@@ -1,6 +1,6 @@
 import { type FC } from "react";
 
-import DataTable from "@/components/common/DataTable";
+import DataTable from "@/components/ui/common/DataTable";
 
 interface LandTableProps {
   paPlayer: {
