--- game/src/pages/api/tickah.ts
@@ -1,11 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-
-export const config = {
-  runtime: "edge",
-};
-
-export default (request: NextRequest) => {
-  return NextResponse.json({
-    name: `Hello, from ${request.url} I'm now an Edge Function!`,
-  });
-};
